Added new talk.py that derives from artifact.py
